set external minification module
